+++ src/auth/authentication-guard.js
@@ -1,6 +1,6 @@
 import { withAuthenticationRequired } from "@auth0/auth0-react";
 import React from "react";
-import { PageLoader } from "../PageLoader";
+import { PageLoader } from "../components/PageLoader";
 
 export const AuthenticationGuard = ({ component }) => {
   const Component = withAuthenticationRequired(component, {
