+++ src/components/PageLayout.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { NavBar } from "./navBar/NavBar";
+import { NavBar } from "../navBar/NavBar";
 
 export const PageLayout = ({ children }) => {
   return (
