+++ src/navBar/NavBarButtons.js
@@ -1,8 +1,8 @@
 import { useAuth0 } from "@auth0/auth0-react";
 import React from "react";
-import { LoginButton } from "../LoginButton";
-import { LogoutButton } from "../LogoutButton";
-import { SignupButton } from "../SignupButton";
+import { LoginButton } from "../buttons/LoginButton";
+import { LogoutButton } from "../buttons/LogoutButton";
+import { SignupButton } from "../buttons/SignupButton";
 
 export const NavBarButtons = () => {
   const { isAuthenticated } = useAuth0();
